% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{cbind_abind}
\alias{cbind_abind}
\title{Combine Results from Parallel Processing}
\usage{
cbind_abind(a, b)
}
\arguments{
\item{a}{First imputation result}

\item{b}{Second imputation result}
}
\value{
Combined results list with:
\item{y_imp}{Combined imputed values}
\item{gibbs_res}{Combined Gibbs sampling results}
}
\description{
Combines imputation results from parallel processing
}
