% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{glmnet_pred}
\alias{glmnet_pred}
\title{Elastic Net Prediction}
\usage{
glmnet_pred(x, y, alpha = 0.5, lambda = 0.01)
}
\arguments{
\item{x}{data.frame containing predictor variables}

\item{y}{numeric vector of response values}

\item{alpha}{mixing parameter between ridge and lasso (default: 0.5)}

\item{lambda}{regularization parameter (default: 0.01)}
}
\value{
numeric vector of predicted values
}
\description{
Fits and predicts using an elastic net model
}
