% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{knn_pred}
\alias{knn_pred}
\title{K-Nearest Neighbors Prediction}
\usage{
knn_pred(x, y)
}
\arguments{
\item{x}{data.frame containing predictor variables}

\item{y}{numeric vector of response values}
}
\value{
numeric vector of predicted values
}
\description{
Fits and predicts using a k-nearest neighbors model (k=5)
}
