% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{kNN_wrapper}
\alias{kNN_wrapper}
\title{k-Nearest Neighbors Imputation Wrapper}
\usage{
kNN_wrapper(data, ...)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}

\item{...}{Additional arguments passed to impute.knn}
}
\value{
Imputed dataset
}
\description{
Wrapper function for k-NN imputation
}
