% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{lm_pred}
\alias{lm_pred}
\title{Prediction Functions for Various Models}
\usage{
lm_pred(x, y)
}
\arguments{
\item{x}{data.frame containing predictor variables}

\item{y}{numeric vector of response values}
}
\value{
numeric vector of predicted values
}
\description{
A collection of wrapper functions for different prediction models

Fits and predicts using a linear regression model
}
