% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{MAR_generate}
\alias{MAR_generate}
\title{Generate Missing at Random (MAR) Data}
\usage{
MAR_generate(data, mis_prop = 0.5)
}
\arguments{
\item{data}{A numeric matrix or data frame}

\item{mis_prop}{Proportion of missing values to generate (between 0 and 1)}
}
\value{
A list containing:
\item{data_res}{The dataset with generated missing values}
\item{mis_idx}{Indices of missing values}
}
\description{
Creates missing values in a dataset completely at random
}
