% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{Mean_wrapper}
\alias{Mean_wrapper}
\title{Mean Imputation Wrapper}
\usage{
Mean_wrapper(data)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}
}
\value{
Imputed dataset with means replacing NA values
}
\description{
Imputes missing values with column means
}
