% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{Median_wrapper}
\alias{Median_wrapper}
\title{Median Imputation Wrapper}
\usage{
Median_wrapper(data)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}
}
\value{
Imputed dataset with medians replacing NA values
}
\description{
Imputes missing values with column medians
}
