% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{MNAR_gen_imp}
\alias{MNAR_gen_imp}
\title{Generate and Impute MNAR Data}
\usage{
MNAR_gen_imp(
  data_c,
  mis_var_prop = seq(0.1, 0.8, 0.1),
  var_mis_prop = seq(0.1, 0.8, 0.1),
  impute_list = c("kNN_wrapper", "SVD_wrapper", "HM_wrapper", "QRILC_wrapper"),
  cores = 5
)
}
\arguments{
\item{data_c}{Complete data matrix}

\item{mis_var_prop}{Vector of missing variable proportions}

\item{var_mis_prop}{Vector of value-based missingness proportions}

\item{impute_list}{Vector of imputation method names}

\item{cores}{Number of cores for parallel processing}
}
\value{
List containing imputation results, complete data, and method names
}
\description{
Generates Missing Not At Random (MNAR) data and performs multiple imputations
}
