% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{MNAR_generate}
\alias{MNAR_generate}
\title{Generate Missing Not at Random (MNAR) Data}
\usage{
MNAR_generate(data_c, mis_var = 0.5, var_prop = seq(0.3, 0.6, 0.1))
}
\arguments{
\item{data_c}{A numeric matrix or data frame}

\item{mis_var}{Proportion of variables to contain missing values or character vector of variable names}

\item{var_prop}{Vector of proportions for value-based missingness cutoffs}
}
\value{
A list containing:
\item{data_mis}{The dataset with generated missing values}
\item{mis_idx_df}{Matrix of indices for missing values}
}
\description{
Creates missing values based on the values themselves
}
