% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVI_global.R
\name{multiplot}
\alias{multiplot}
\title{Create Multiple ggplot2 Plots in a Single Layout}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot2 objects}

\item{plotlist}{Optional list of ggplot2 objects}

\item{file}{Output file (optional)}

\item{cols}{Number of columns in the layout}

\item{layout}{Optional matrix specifying the layout design}
}
\description{
Combines multiple ggplot2 plots into a single layout
}
