% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{NRMSE_list}
\alias{NRMSE_list}
\title{Calculate NRMSE Between Multiple Imputed Datasets}
\usage{
NRMSE_list(imp_list, miss_df, method_names)
}
\arguments{
\item{imp_list}{List of imputed datasets}

\item{miss_df}{Original dataset with missing values}

\item{method_names}{Vector of method names}
}
\value{
Matrix of pairwise NRMSE values
}
\description{
Calculates pairwise NRMSE between multiple imputed datasets
}
