% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{NRMSE_rank_cal_plot}
\alias{NRMSE_rank_cal_plot}
\title{Calculate and Plot NRMSE Rankings}
\usage{
NRMSE_rank_cal_plot(
  impute_results,
  plot = T,
  x = "Miss_Prop",
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{Results from MNAR_gen_imp}

\item{plot}{Logical, whether to create plot}

\item{x}{Variable for x-axis ('Miss_Prop' or 'Miss_Num')}

\item{colors}{Optional color vector for plotting}

\item{shapes}{Optional shape vector for plotting}
}
\value{
List containing NRMSE rankings and melted data frame for plotting
}
\description{
Calculates and visualizes NRMSE rankings for different imputation methods
}
