% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{PCA_pro_list}
\alias{PCA_pro_list}
\title{Calculate Procrustes Analysis for Multiple Imputed Datasets}
\usage{
PCA_pro_list(imp_list, nPCs = 2, method_names)
}
\arguments{
\item{imp_list}{List of imputed datasets}

\item{nPCs}{Number of principal components}

\item{method_names}{Vector of method names}
}
\value{
List containing Procrustes sum of squares matrix and detailed results
}
\description{
Performs pairwise Procrustes analysis on PCA results from multiple imputed datasets
}
