% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{Procrustes_cal_plot}
\alias{Procrustes_cal_plot}
\title{Calculate and Plot Procrustes Analysis}
\usage{
Procrustes_cal_plot(
  impute_results,
  DR = "PCA",
  lg = F,
  nPCs = 2,
  outcome = NULL,
  x = "Miss_Prop",
  plot = T,
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{Results from MNAR_gen_imp}

\item{DR}{Dimensionality reduction method ('PCA' or 'PLS')}

\item{lg}{Logical, whether to log transform data}

\item{nPCs}{Number of principal components}

\item{outcome}{Outcome variable for PLS}

\item{x}{Variable for x-axis}

\item{plot}{Logical, whether to create plot}

\item{colors}{Optional color vector for plotting}

\item{shapes}{Optional shape vector for plotting}
}
\value{
List containing Procrustes sum of squares and melted data frame for plotting
}
\description{
Performs Procrustes analysis on dimensionality reduction results
}
