% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{QRILC_wrapper}
\alias{QRILC_wrapper}
\title{Quantile Regression Imputation of Left-Censored Data Wrapper}
\usage{
QRILC_wrapper(data, ...)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}

\item{...}{Additional arguments passed to impute.QRILC}
}
\value{
Imputed dataset
}
\description{
Wrapper function for QRILC imputation with log transformation
}
