% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{rf_pred}
\alias{rf_pred}
\title{Random Forest Prediction}
\usage{
rf_pred(x, y, ntree = 200, ...)
}
\arguments{
\item{x}{data.frame containing predictor variables}

\item{y}{numeric vector of response values}

\item{ntree}{number of trees to grow (default: 200)}

\item{...}{additional parameters passed to randomForest}
}
\value{
numeric vector of predicted values
}
\description{
Fits and predicts using a random forest model
}
