% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{RF_wrapper}
\alias{RF_wrapper}
\title{Random Forest Imputation Wrapper}
\usage{
RF_wrapper(data, ...)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}

\item{...}{Additional arguments passed to missForest}
}
\value{
Imputed dataset
}
\description{
Wrapper function for missForest imputation
}
