% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{Generate Samples from Truncated Normal Distribution}
\usage{
rnorm_trunc(n, mu, std, lo = -Inf, hi = Inf)
}
\arguments{
\item{n}{Number of samples to draw}

\item{mu}{Mean of the normal distribution}

\item{std}{Standard deviation of the normal distribution}

\item{lo}{Lower bound (default -Inf)}

\item{hi}{Upper bound (default Inf)}
}
\value{
Vector of n samples from the truncated normal distribution
}
\description{
Draws n samples from a truncated normal distribution N(mu, std^2|\link{lo, hi})
}
