% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_funcs.R
\name{rpart_pred}
\alias{rpart_pred}
\title{Decision Tree Prediction}
\usage{
rpart_pred(x, y)
}
\arguments{
\item{x}{data.frame containing predictor variables}

\item{y}{numeric vector of response values}
}
\value{
numeric vector of predicted values
}
\description{
Fits and predicts using a decision tree model
}
