% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{single_impute_iters}
\alias{single_impute_iters}
\title{Single Variable Gibbs Sampler Imputation}
\usage{
single_impute_iters(
  x,
  y,
  y_miss,
  y_real = NULL,
  imp_model = "glmnet_pred",
  lo = -Inf,
  hi = Inf,
  iters_each = 100,
  gibbs = c()
)
}
\arguments{
\item{x}{Matrix of predictor variables}

\item{y}{Response variable with missing values}

\item{y_miss}{Original response with missing values}

\item{y_real}{True values (optional, for testing)}

\item{imp_model}{Imputation model function name}

\item{lo}{Lower bound for truncated normal}

\item{hi}{Upper bound for truncated normal}

\item{iters_each}{Number of iterations}

\item{gibbs}{Indices for Gibbs sampling monitoring}
}
\value{
List containing:
\item{y_imp}{Imputed values}
\item{gibbs_res}{Gibbs sampling results array}
}
\description{
Performs Gibbs sampling imputation for a single variable
}
