% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{SVD_wrapper}
\alias{SVD_wrapper}
\title{Singular Value Decomposition Imputation Wrapper}
\usage{
SVD_wrapper(data, K = 5)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}

\item{K}{Number of components for SVD imputation}
}
\value{
Imputed dataset
}
\description{
Wrapper function for SVD-based imputation with scaling
}
