% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_evaluation.R
\name{Ttest_cor_cal_plot}
\alias{Ttest_cor_cal_plot}
\title{Calculate and Plot T-test Correlations}
\usage{
Ttest_cor_cal_plot(
  impute_results,
  group = NULL,
  plot = T,
  x = "Miss_Prop",
  cor = "P",
  colors = NULL,
  shapes = NULL
)
}
\arguments{
\item{impute_results}{Results from MNAR_gen_imp}

\item{group}{Group labels for t-test}

\item{plot}{Logical, whether to create plot}

\item{x}{Variable for x-axis}

\item{cor}{Correlation type ('P' for Pearson, 'S' for Spearman)}

\item{colors}{Optional color vector for plotting}

\item{shapes}{Optional shape vector for plotting}
}
\value{
List containing correlation results and melted data frames
}
\description{
Calculates correlations between t-test p-values from complete and imputed data
}
