% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_wrapper.R
\name{Zero_wrapper}
\alias{Zero_wrapper}
\title{Zero Imputation Wrapper}
\usage{
Zero_wrapper(data)
}
\arguments{
\item{data}{A numeric matrix or data frame with missing values}
}
\value{
Imputed dataset with zeros replacing NA values
}
\description{
Imputes missing values with zeros
}
